EMS-693: Fied validation rules changes and text
