+++ src/ui/server/controllers/insurance/your-buyer/validation/index.ts
@@ -1,6 +1,6 @@
 import { RequestBody, ValidationErrors } from '../../../../../types';
 import combineValidationRules from '../../../../helpers/combine-validation-rules';
-import validationRules from './rules/rules';
+import validationRules from './rules';
 
 const yourBuyerDetailsValidation = (formBody: RequestBody): ValidationErrors => combineValidationRules(validationRules, formBody);
