+++ src/ui/server/controllers/insurance/your-buyer/validation/rules/index.ts
@@ -1,0 +1,7 @@
+import { countryValidationRules } from './countryValidationRules';
+import { ValidationErrors } from '../../../../../../types';
+import { organisationNameValidationRules } from './organisationNameValidationRules';
+
+const rules = [countryValidationRules, organisationNameValidationRules] as Array<() => ValidationErrors>;
+
+export default rules;
