--- src/ui/server/controllers/insurance/your-buyer/validation/rules/rules.ts
@@ -1,6 +1,0 @@
-import { validationRules } from './formValidationRules';
-import { ValidationErrors } from '../../../../../../types';
-
-const rules = [validationRules] as Array<() => ValidationErrors>;
-
-export default rules;
